refactor: move purchaseSmelter to its own file
